QF3: removed some more warnings!
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3985 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/basicimageimporters/qfimagereaderraw.h
@@ -118,7 +118,7 @@ class QFImageReaderRAW: public QFImporterImageSeries
         /** \brief move on to the next frame in the file. return \c false if no further image exists */
         inline virtual bool nextFrame() {
             if (rawFile) {
-                if (rawFile->read((char*)currentFrame, width*height*channels*uint64_t(sizeof(T)))==width*height*channels*uint64_t(sizeof(T))) {
+                if ((uint64_t)rawFile->read((char*)currentFrame, width*height*channels*uint64_t(sizeof(T)))==width*height*channels*uint64_t(sizeof(T))) {
                     frame++;
                     return true;
                 }
